+++ tests/dynamic_data/publications/wrong_input/wrong_input_test.py
@@ -1570,5 +1570,5 @@ class TestPublication(base_test.TestSingleRestPublication):
                                                                      name=publication.name,
                                                                      rest_publication_detail=rest_publication_detail,
                                                                      headers=None,
-                                                                     different_value_keys=['file', 'geodata_type'],
+                                                                     different_value_keys=['geodata_type'],
                                                                      )